Created a folder to check best logging method
The folder contains a Runner and a MyTestClass.
The MyTestClass class contains a method that throws a new Exception.
The runner creates an instance of MyTestClass and calls the error method.
The thing to note is how the error log is printed on the log file with
and without using the Gson.toJson() method.